pyproject.toml
[project]
name = "fit_agent"
version = "0.1.0"
description = "fit-agent"
requires-python = ">=3.11"

dependencies = [
    "aegra-cli>=0.6.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.9",
    "pydub>=0.25.0",
    "audioop-lts>=0.2.0; python_version>='3.13'",
    "langgraph>=1.0.8",
    "langchain>=0.3.14",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.12",
    "langchain-community>=0.3.0",
    "httpx-sse>=0.4.0",
    "PyJWT>=2.0.0",
    "alembic>=1.14.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "psycopg[binary,pool]>=3.0.0",
    "passlib[bcrypt]>=1.7.4",
    "email-validator>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fit_agent"]
